mega code refactorization
