+++ .gitignore
@@ -5,3 +5,5 @@ __pycache__
 __pycache__/KNN.cpython-312.pyc
 CIFAR_100_incremental_T_5_table_1
 CIFAR_100_incremental_T_5_table_1.7z
+
+venv
