+++ results/old/kmeans_vit.txt
