+++ results/old/kmeans_vit_2.txt
