+++ results/old/knn_resnet.txt
