+++ results/old/knn_vit.txt
